+++ src/controllers/ElementIndexesController.php
@@ -19,6 +19,7 @@ use craft\elements\conditions\ElementConditionRuleInterface;
 use craft\elements\db\ElementQueryInterface;
 use craft\elements\exporters\Raw;
 use craft\events\ElementActionEvent;
+use craft\helpers\Cp;
 use craft\helpers\ElementHelper;
 use craft\services\ElementSources;
 use yii\base\InvalidValueException;
@@ -800,5 +801,65 @@ class ElementIndexesController extends BaseElementsController
         }
 
         return $exporterData;
+    }
+
+    /**
+     * Returns the updated table attribute HTML for an element.
+     *
+     * @return Response
+     * @throws BadRequestHttpException
+     * @throws ForbiddenHttpException
+     */
+    public function actionElementTableHtml(): Response
+    {
+        $this->requireAcceptsJson();
+
+        if (!$this->sourceKey) {
+            throw new BadRequestHttpException("Request missing required body param");
+        }
+
+        /** @var string|ElementInterface $elementType */
+        $elementType = $this->elementType;
+        $id = $this->request->getRequiredBodyParam('id');
+        $siteId = $this->request->getRequiredBodyParam('siteId');
+        $site = $siteId ? Craft::$app->getSites()->getSiteById($siteId) : null;
+
+        if (!$id || !is_numeric($id)) {
+            throw new BadRequestHttpException("Invalid element ID: $id");
+        }
+
+        if (!$site) {
+            throw new BadRequestHttpException("Invalid site ID: $siteId");
+        }
+
+        if (Craft::$app->getIsMultiSite() && !Craft::$app->getUser()->checkPermission("editSite:$site->uid")) {
+            throw new ForbiddenHttpException('User not authorized to edit content for this site.');
+        }
+
+        /** @var ElementInterface|null $element */
+        $element = $elementType::find()
+            ->id($id)
+            ->drafts(null)
+            ->provisionalDrafts(null)
+            ->revisions(null)
+            ->siteId($siteId)
+            ->status(null)
+            ->one();
+
+        if (!$element) {
+            throw new BadRequestHttpException("Invalid element ID: $id");
+        }
+
+        $attributes = Craft::$app->getElementSources()->getTableAttributes($this->elementType, $this->sourceKey);
+        $attributeHtml = [];
+
+        foreach ($attributes as [$attribute]) {
+            $attributeHtml[$attribute] = $element->getTableAttributeHtml($attribute);
+        }
+
+        return $this->asJson([
+            'elementHtml' => Cp::elementHtml($element, $this->context),
+            'attributeHtml' => $attributeHtml,
+        ]);
     }
 }