+++ src/web/assets/cp/src/js/BaseElementSelectInput.js
@@ -300,7 +300,6 @@ Craft.BaseElementSelectInput = Garnish.Base.extend(
     createElementEditor: function ($element, settings) {
       settings = Object.assign(
         {
-          elementSelectInput: this,
           prevalidate: this.settings.prevalidate,
         },
         settings